[NEW] add refreshPollFields in pollUtils.js
